infrastructure/apps/janitor/requirements.txt
atlassian-python-api
requests
pg8000
